Try to modify the MessageClient using raw socket
But now it is not finished. I think it is better to roll back to the
old version with the frame of Netty.
Now I start to think it might be a stupid idea to change the code back
to the raw socket version. We will see. Now I just push and branch the
current code as “Dev-A-Try-To-Raw-Socket”